refactor(opencode): split message event handlers
